Creating sanitychecks module
