+++ pronto/static/modules/sanitycheck/settings.js
@@ -2,8 +2,12 @@ import {finaliseHeader} from "../../header.js"
 import * as ui from "../../ui.js";
 
 
-
 $(function () {
     finaliseHeader();
+
+    $('.ui.tabular.menu .item').tab({
+        auto: true,
+        path: '/api'
+    });
 
 });
