+++ pronto/templates/sanitycheck/settings.html
@@ -14,6 +14,16 @@
 <div class="ui container basic vertical segment">
     <h1 class="ui dividing header">Sanity check settings</h1>
 
+    <div class="ui top attached tabular menu">
+        <div class="item" data-tab="tab-name">Tab Name</div>
+        <div class="item" data-tab="tab-name2">Tab Name 2</div>
+    </div>
+    <div class="ui tab" data-tab="tab-name">
+        <!-- Tab Content !-->
+    </div>
+    <div class="ui tab" data-tab="tab-name2">
+        <!-- Tab Content !-->
+    </div>
 </div>
 
 {% include 'footer.html' %}
